changed data type

+++ src/main/java/org/example/Addition.java
@@ -2,7 +2,7 @@ package org.example;
 
 public class Addition {
 
-    public int add(int a, int b) {
+    public float add(float a, float b) {
         return a + b;
     }
 }
